Move the custom inflector config into the config folder
This isn't a model class so it's more logical to house it in the config folder instead.
